+++ src/config/acceptValues.js
@@ -1,7 +1,7 @@
 const acceptNumberAndLetter = '^[a-zA-Z0-9_.-]*$';
 // const acceptLeague = ['NBA', 'CBA', 'MLB', 'KBO', 'NPB', 'CPBL', 'eSoccer'];
 
-const acceptLeague = ['NBA', 'MLB', 'KBO', 'NPB', 'CPBL', 'eSoccer'];
+const acceptLeague = ['NBA', 'MLB', 'KBO', 'NPB', 'CPBL', 'eSoccer', 'Soccer'];
 module.exports = {
   acceptNumberAndLetter,
   acceptLeague